% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fool_proofing.r
\name{check_data}
\alias{check_data}
\title{Check Data for Missing Response Values}
\usage{
check_data(y_names, data)
}
\arguments{
\item{y_names}{A vector of column names for the outcome variable(s).}

\item{data}{A data frame containing the data.}
}
\value{
A data frame with response indicator column `r`.
}
\description{
This method checks if the outcome variable has missing values and creates a response indicator column (`r`) if necessary.
If the outcome variable does not have missing values, it checks if the response indicator column (`r`) already exists.
}
\examples{
\dontrun{
data = data.frame(outcome = c(1, NA, 3, 4, NA), covariate = c(2, 3, 4, 5, 6))
y_names = "outcome"
cleaned_data = ebmr_algorithm$private$check_data(y_names, data)
}

}
\keyword{internal}
