% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_methods.r
\name{EBMR_IPW_with_locally_misspecified_model}
\alias{EBMR_IPW_with_locally_misspecified_model}
\title{Compute the IPW Estimator with Locally Misspecified Model}
\usage{
EBMR_IPW_with_locally_misspecified_model(
  ps.matrix,
  perturb_ps,
  exp_tilt,
  exp_tilt_x_names,
  h_x_names,
  nu_init = rep(1/J, J),
  se.fit = FALSE
)
}
\arguments{
\item{ps.matrix}{A matrix of estimated propensity scores for each observation and model.}

\item{perturb_ps}{An integer specifying the index of the propensity score model to be perturbed.
This model is typically assumed to be the well-specified model among the candidate models.}

\item{exp_tilt}{A function defining the exponential tilt model.}

\item{exp_tilt_x_names}{A character vector specifying the covariates used in the exponential tilt model.}
}
\value{
A list containing the following elements:
\describe{
\item{\code{mu_ipw}}{The IPW estimator for the population mean \eqn{\mu_0} using the perturbed propensity scores.}
\item{\code{se_ipw}}{The standard error of the IPW estimator.}
\item{\code{ps.matrix}}{The updated matrix of estimated propensity scores after perturbation.}
\item{\code{nu.hat}}{The estimated coefficients \eqn{\nu} from the ensemble method.}
\item{\code{w.hat}}{The estimated weights for the propensity score models.}
\item{\code{imbalance}}{A measure of imbalance provided by the ensemble propensity scores.}
}
}
\description{
This function computes the Inverse Probability Weighting (IPW) estimator for
the population mean \eqn{\mu_0}, incorporating a sensitivity analysis by
perturbing a specified propensity score model using an exponential tilt model.
The method allows users to assess the impact of local model misspecification
on the estimator.
}
\details{
The function perturbs the designated propensity score model using an exponential tilt model,
which is a sensitivity analysis technique for assessing local model misspecification. The
impact of perturbation on the IPW estimator is then examined. The ensemble method is used
to estimate the optimal weights for the propensity score models, ensuring balance across covariates.
}
\examples{
\dontrun{
ebmr <- EBMRAlgorithm$new(data = data)
ipw_sensitivity <- ebmr$EBMR_IPW_with_locally_misspecified_model(
  ps.matrix = estimated_ps_matrix,
  perturb_ps = 2,
  exp_tilt = function(y, data) exp(y * data$covariate),
  exp_tilt_x_names = c("covariate")
)
print(ipw_sensitivity)
}
}
