% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBMRAlgorithm.R
\name{EBMRAlgorithm_perturb}
\alias{EBMRAlgorithm_perturb}
\title{EBMRAlgorithm Class}
\description{
The \code{EBMRAlgorithm_perturb} class provides an implementation for estimating
the coefficients in a propensity score model and related functionalities
based on the ensemble method for inverse probability weighting (IPW).
This class allows users to perform propensity score fitting and estimation
for the specified data and propensity score specifications.
}
\section{Public Methods}{

The following methods are available in the EBMRAlgorithm class:

\code{initialize(y_names, ps_specifications, data)}:
Initializes the class and estimates the propensity scores based on the provided formulas,
covariates to be balanced, inverse link function, and the specified data.

\code{WangShaoKim2014(formula, h_x_names, inv_link, init = NULL)}:
Implements the Wang, Shao, and Kim (2014) approach for estimating propensity score models in the presence of missing-not-at-random (MNAR) data.

\code{EBMR_IPW(h_x_names, true_ps = NULL)}:
Computes the Inverse Probability Weighting (IPW) estimator for the population mean \eqn{\mu_0},
using the propensity scores estimated by the ensemble method. It also computes standard errors and other related
quantities for the estimator, including the estimator when the true propensity score is provided.
}

\section{Private Fields}{

\describe{
\item{\code{r}}{The response indicator.}
\item{\code{y}}{The outcome variable(s) used in the model.}
\item{\code{n}}{The number of observations.}
\item{\code{ps_fit.list}}{List of fitted propensity score models.}
}
}

\examples{
\dontrun{
# Define propensity score specifications (for illustration)
ps_specifications = list(
  formula.list = list(formula1, formula2),
  h_x_names.list = list(h_x1, h_x2),
  inv_link = inv_link_function
)

# Initialize the EBMRAlgorithm class
ebmr = EBMRAlgorithm$new(y_names = "outcome", ps_specifications = ps_specifications, data = dataset)
}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EBMRAlgorithm_perturb-new}{\code{EBMRAlgorithm_perturb$new()}}
\item \href{#method-EBMRAlgorithm_perturb-EBMR_IPW_perturb}{\code{EBMRAlgorithm_perturb$EBMR_IPW_perturb()}}
\item \href{#method-EBMRAlgorithm_perturb-EBMR_IPW_with_locally_misspecified_model}{\code{EBMRAlgorithm_perturb$EBMR_IPW_with_locally_misspecified_model()}}
\item \href{#method-EBMRAlgorithm_perturb-WangShaoKim2014_perturb}{\code{EBMRAlgorithm_perturb$WangShaoKim2014_perturb()}}
\item \href{#method-EBMRAlgorithm_perturb-clone}{\code{EBMRAlgorithm_perturb$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EBMRAlgorithm_perturb-new"></a>}}
\if{latex}{\out{\hypertarget{method-EBMRAlgorithm_perturb-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EBMRAlgorithm_perturb$new(y_names, ps_specifications, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EBMRAlgorithm_perturb-EBMR_IPW_perturb"></a>}}
\if{latex}{\out{\hypertarget{method-EBMRAlgorithm_perturb-EBMR_IPW_perturb}{}}}
\subsection{Method \code{EBMR_IPW_perturb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EBMRAlgorithm_perturb$EBMR_IPW_perturb(h_x_names, true_ps = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EBMRAlgorithm_perturb-EBMR_IPW_with_locally_misspecified_model"></a>}}
\if{latex}{\out{\hypertarget{method-EBMRAlgorithm_perturb-EBMR_IPW_with_locally_misspecified_model}{}}}
\subsection{Method \code{EBMR_IPW_with_locally_misspecified_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EBMRAlgorithm_perturb$EBMR_IPW_with_locally_misspecified_model(
  ps.matrix,
  perturb_ps,
  exp_tilt,
  exp_tilt_x_names,
  h_x_names
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EBMRAlgorithm_perturb-WangShaoKim2014_perturb"></a>}}
\if{latex}{\out{\hypertarget{method-EBMRAlgorithm_perturb-WangShaoKim2014_perturb}{}}}
\subsection{Method \code{WangShaoKim2014_perturb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EBMRAlgorithm_perturb$WangShaoKim2014_perturb(
  formula,
  h_x_names,
  inv_link,
  init = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EBMRAlgorithm_perturb-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EBMRAlgorithm_perturb-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EBMRAlgorithm_perturb$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
