% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessor.r
\name{separate_variable_types}
\alias{separate_variable_types}
\title{Separate Continuous and Discrete Variables from a Data Frame}
\usage{
separate_variable_types(x)
}
\arguments{
\item{x}{A data frame containing the variables to be separated.}
}
\value{
A list containing:
\describe{
  \item{\code{x1}}{A data frame with the discrete variables.}
  \item{\code{x2}}{A data frame with the continuous variables.}
  \item{\code{x1_names}}{A character vector with the names of the discrete variables.}
  \item{\code{x2_names}}{A character vector with the names of the continuous variables.}
}
}
\description{
This method separates continuous and discrete variables from a given data frame
based on their characteristics. Continuous variables are numeric columns with more than 10 unique values,
while discrete variables are either non-numeric or numeric columns with fewer than or equal to 10 unique values.
}
\details{
The function separates variables into two categories:
\itemize{
  \item Continuous variables: Numeric variables with more than 10 unique values.
  \item Discrete variables: Non-numeric variables or numeric variables with 10 or fewer unique values.
}
The function returns the separated variables in two data frames: one for continuous variables (`x2`)
and one for discrete variables (`x1`), along with the corresponding names of these variables.
}
\examples{
\dontrun{
data <- data.frame(
  age = c(25, 30, 35, 40),
  gender = c('Male', 'Female', 'Female', 'Male'),
  salary = c(50000, 60000, 55000, 70000)
)
ebmr <- EBMRAlgorithm$new(data = data)
result <- ebmr$separate_variable_types(data)
print(result)
}

}
\keyword{internal}
