% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WangShaoKim2014.r
\name{WangShaoKim2014}
\alias{WangShaoKim2014}
\title{Wang, Shao, and Kim (2014) Method for Propensity Score Estimation}
\usage{
WangShaoKim2014(formula, h_x_names, inv_link, init = NULL)
}
\arguments{
\item{formula}{A formula specifying the relationship between the response and predictors.}

\item{h_x_names}{A character vector of variable names to be balanced.}

\item{inv_link}{An inverse link function applied to the linear predictor.}

\item{init}{(Optional) A numeric vector specifying the initial values for the model parameters. Defaults to `NULL`.}
}
\value{

}
\description{
Implements the Wang, Shao, and Kim (2014) approach for estimating propensity score models in the presence of missing-not-at-random (MNAR) data.
}
\examples{
\dontrun{
data <- data.frame(x = rnorm(100), y = rnorm(100), r = rbinom(100, 1, 0.5))
ebmr <- EBMRAlgorithm$new(data = data)
result <- ebmr$WangShaoKim2014(
  formula = r ~ x,
  h_x_names = c("x"),
  inv_link = function(eta) 1 / (1 + exp(-eta))
)
print(result$coefficients)
}

}
\references{
Wang, Shao, & Kim (2014). "An instrumental variable approach for identification and estimation with nonignorable nonresponse."
}
